+++ frontend/src/services/map/markerGeneratorService.js
