+++ frontend/src/services/map/markerManagerService.js
